first version of tests that reproduce mem leak

+++ python/arcticdb/util/test.py
@@ -299,9 +299,8 @@ def random_integers(size, dtype):
     platform_int_info = np.iinfo("int_")
     iinfo = np.iinfo(dtype)
     return np.random.randint(
-        max(iinfo.min, platform_int_info.min), min(iinfo.max, platform_int_info.max), size=size
-    ).astype(dtype)
-
+        max(iinfo.min, platform_int_info.min), min(iinfo.max, platform_int_info.max), size=size, dtype=dtype
+    )
 
 def get_wide_dataframe(size=10000, seed=0):
     np.random.seed(seed)
@@ -321,7 +320,6 @@ def get_wide_dataframe(size=10000, seed=0):
             "bool": np.random.randn(size) > 0,
         }
     )
-
 
 def get_pickle():
     return (
